[#7]:svarga:pkg, refactor into Julia package with test and example
