+++ Project.toml
@@ -1,0 +1,14 @@
+name = "TinyCrypto"
+uuid = "a4b5da31-a86c-473b-96f4-ea8fbe6abb4b"
+authors = ["svarga <steven@vargaconsulting.ca>"]
+version = "0.1.0"
+
+[deps]
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[compat]
+LinearAlgebra = "1.11.0"
+Random = "1.11.0"
+Test = "1.11.0"
