+++ src/TinyCrypto.jl
@@ -1,0 +1,5 @@
+module TinyCrypto
+
+include("curves/curve.jl")
+
+end
