+++ src/curves/curve.jl
